pyproject.toml
[project]
name = "langchain-langgraph-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-deepseek>=1.0.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
]
